+++ packages/xchain-ethereum/src/client.ts
@@ -107,7 +107,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * @param {EthereumClientParams} params
    */
   constructor({
-    network = Network.Testnet,
+    network = Network.Mainnet,
     feeBounds = {
       lower: LOWER_FEE_BOUND,
       upper: UPPER_FEE_BOUND,