+++ packages/xchain-thorchain-amm/src/thorchain-amm.ts
@@ -1,9 +1,11 @@
 import {
   AddliquidityPosition,
+  EstimateAddLP,
   EstimateSwapParams,
-  RemoveLiquidityPosition,
+  EstimateWithdrawLP,
   ThorchainQuery,
   TxDetails,
+  WithdrawLiquidityPosition,
 } from '@xchainjs/xchain-thorchain-query'
 import { BigNumber } from 'bignumber.js'
 
@@ -94,6 +96,25 @@ export class ThorchainAMM {
       waitTimeSeconds,
     })
   }
+
+  /**
+   * Wraps estimate from thorchain query
+   * @param params - estimate add liquidity
+   * @returns - Estimate add lp object
+   */
+  public async estimateAddLiquidity(params: AddliquidityPosition): Promise<EstimateAddLP> {
+    return this.thorchainQuery.estimateAddLP(params)
+  }
+
+  /**
+   * Wraps estimate withdraw from thorchain query
+   * @param params - estimate withdraw liquidity
+   * @returns - Estimate withdraw lp object
+   */
+  public async estimateWithdrawLiquidity(params: WithdrawLiquidityPosition): Promise<EstimateWithdrawLP> {
+    return this.thorchainQuery.estimateWithdrawLP(params)
+  }
+
   /**
    *
    * @param wallet - wallet class
@@ -103,11 +124,12 @@ export class ThorchainAMM {
   public async addLiquidityPosition(wallet: Wallet, params: AddliquidityPosition): Promise<TxSubmitted[]> {
     // Check amounts are greater than fees and use return estimated wait
     const checkLPAdd = await this.thorchainQuery.estimateAddLP(params)
-
+    if (!checkLPAdd.canAdd) throw Error(`${checkLPAdd.errors}`)
     return wallet.addLiquidity({
       asset: params.asset,
       rune: params.rune,
       waitTimeSeconds: checkLPAdd.estimatedWaitSeconds,
+      assetPool: checkLPAdd.assetPool,
     })
   }
   /**
@@ -116,14 +138,17 @@ export class ThorchainAMM {
    * @param wallet - wallet needed to perform tx
    * @return
    */
-  public async removeLiquidityPosition(wallet: Wallet, params: RemoveLiquidityPosition): Promise<TxSubmitted[]> {
+  public async withdrawLiquidityPosition(wallet: Wallet, params: WithdrawLiquidityPosition): Promise<TxSubmitted[]> {
     // Caution Dust Limits: BTC,BCH,LTC chains 10k sats; DOGE 1m Sats; ETH 0 wei; THOR 0 RUNE.
-    const estimateWithrawLp = await this.thorchainQuery.estimateWithdrawLP(params)
-    return wallet.removeLiquidity({
-      asset: estimateWithrawLp.transactionFee.assetFee,
-      rune: estimateWithrawLp.transactionFee.runeFee,
+    const withdrawParams = await this.thorchainQuery.estimateWithdrawLP(params)
+    return wallet.withdrawLiquidity({
+      assetFee: withdrawParams.transactionFee.assetFee,
+      runeFee: withdrawParams.transactionFee.runeFee,
+      waitTimeSeconds: withdrawParams.estimatedWaitSeconds,
       percentage: params.percentage,
-      waitTimeSeconds: estimateWithrawLp.estimatedWaitSeconds,
+      assetPool: withdrawParams.assetPool,
+      assetAddress: withdrawParams.assetAddress,
+      runeAddress: withdrawParams.runeAddress,
     })
   }
 }