+++ packages/xchain-thorchain-query/src/types.ts
@@ -62,8 +62,8 @@ export type SwapOutput = {
 }
 
 export type UnitData = {
-  liquidityUnits: BaseAmount
-  totalUnits: BaseAmount
+  liquidityUnits: BigNumber
+  totalUnits: BigNumber
 }
 
 export type LiquidityData = {
@@ -132,8 +132,8 @@ export type TxStageStatus = {
 }
 
 export type LiquidityToAdd = {
-  asset: BaseAmount
-  rune: BaseAmount
+  asset: CryptoAmount
+  rune: CryptoAmount
 }
 export type PostionDepositValue = {
   asset: BaseAmount
@@ -146,6 +146,7 @@ export type PoolShareDetail = {
 }
 
 export type EstimateAddLP = {
+  assetPool: string
   slipPercent: BigNumber
   poolShare: PoolShareDetail
   lpUnits: BaseAmount
@@ -156,12 +157,15 @@ export type EstimateAddLP = {
   canAdd: boolean
 }
 export type EstimateWithdrawLP = {
+  assetAddress?: string
+  runeAddress?: string
   slipPercent: BigNumber
   transactionFee: LPFees
   assetAmount: CryptoAmount
   runeAmount: CryptoAmount
   impermanentLossProtection: ILProtectionData
   estimatedWaitSeconds: number
+  assetPool: string
 }
 
 export type LPFees = {
@@ -177,7 +181,7 @@ export type AddliquidityPosition = {
   asset: CryptoAmount
   rune: CryptoAmount
 }
-export type RemoveLiquidityPosition = {
+export type WithdrawLiquidityPosition = {
   asset: Asset
   percentage: number
   assetAddress?: string
@@ -193,16 +197,3 @@ export type PoolRatios = {
   assetToRune: BigNumber
   runeToAsset: BigNumber
 }
-// export type LiquidityProvider = {
-//   asset: string
-//   rune_address: string
-//   asset_address: string
-//   last_add_height: number
-//   last_withdraw_height: number
-//   units: number
-//   pending_rune: number
-//   pending_asset: number
-//   pending_tx_Id: string
-//   rune_deposit_value: number
-//   asset_deposit_value: number
-// }