+++ packages/xchain-thorchain/src/types/client-types.ts
@@ -22,9 +22,9 @@ export type ChainId = string
 export type ChainIds = Record<Network, ChainId>
 
 export type ThorchainClientParams = {
-  clientUrl: ClientUrl
+  clientUrl?: ClientUrl
   explorerUrls?: ExplorerUrls
-  chainIds: ChainIds
+  chainIds?: ChainIds
 }
 
 export type DepositParam = {