+++ packages/xchain-litecoin/CHANGELOG.md
@@ -1,4 +1,11 @@
-# v.0.10.3 (2022-xx-xx)
+# v.10.4 (2022-10-14)
+
+## Update
+
+- Set Default network to `Network.Mainnet`
+- change contructor to accept NodeUrls = Record<Network, string>
+
+# v.0.10.3 (2022-10-04)
 
 ## Update